requirement.txt
asgiref==3.8.1
Django==5.1
django-autoslug==1.9.9
django-autoslug-field==0.2.3
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1